GLODROID: Suppress uevent messages from cooling devices
During suspending when BIG cpu cluster completely went offline the
kernel unregisters cooling device which sends "REMOVE" uevent to all
subscribers.

In case one of the userspace subscribers sets the EPOLLWAKEUP flag, a
wakeup event will be triggered that will cause suspend to abort.

This fixes the suspend feature for the PPP.

Signed-off-by: Roman Stratiienko <r.stratiienko@gmail.com>